// State Pattern feat: The State pattern allows an object to change its behaviour based on its internal state.
This pattern is implemented by having
a context (Canvas class) that can have different states
(the TiledLayer implementations). The context delegates its
behaviour to the current state and can change it at runtime.

In the program, the Canvas class initializes two different states
(MapLayer and Map2Layer) and switches between them when the current
state is done displaying on the canvas.